pyproject.toml
[tool.poetry]
name = "file-retriever"
version = "0.1.0"
description = "BookOps file retriever for FTP/SFTP servers"
authors = ["Charlotte Kostelic <charlottekostelic@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
paramiko = "^3.4.0"
types-paramiko = "^3.4.0.20240423"
types-pyyaml = "^6.0.12.20240311"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "livetest: mark a test as hitting a live ftp/sftp server",
    "tmpdir: mark a test as using a temporary directory",
]


[tool.coverage.run]
relative_files = true
source = ["."]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
